Fix: Users incorrectly associated with 'local' Slurm account
rocks sync users was being invoked before the slurm.sh sub-action,
causing the user to be associated with the default 'local' Slurm
account.

Also, now enumerating Slurm accounts matching the project name to be
associated with the user in lieu of cpu_hours and gpu_hours
conditionals (which had been flawed from start, as the variables are not
defined for user_add).